conftest: func-scoped versions of existing fixtures
existing fixtures allow creating a test container per test class.
this means one container, and its complete state, is shared across
all functions in a test class.

this is ok-ish if that's the idea - test funcs must be ordered in a
way that respects modifications made in previous funcs. this can get
out of control quickly if new funcs are added to a class.

a more convenient pattern is to have the container created per test
func - this way each scenario starts with a clean slate, so:

extend the set of fixtures with their function-scoped versions.

Issues: 4150

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>